Fix build #125: env guard mode_t cast + LicenseStore canonicalString type-check
- AorusEnvGuard: S_IFMT/S_IFLNK are mode_t (UInt16) in this SDK; compare via UInt32
  casts instead of Int32 (fixes "cannot convert Int32 to UInt16").
- LicenseStore.canonicalString: split the big array literal into explicit appends
  (fixes "compiler unable to type-check this expression in reasonable time").

+++ AorusGram/Sources/Features/Subscription/AorusEnvGuard.swift
@@ -167,7 +167,7 @@ enum AorusEnvGuard {
                      "/usr/share", "/var/stash"]
         for path in links {
             var st = stat()
-            if lstat(path, &st) == 0 && (Int32(st.st_mode) & S_IFMT) == S_IFLNK { return true }
+            if lstat(path, &st) == 0 && (UInt32(st.st_mode) & UInt32(S_IFMT)) == UInt32(S_IFLNK) { return true }
         }
         return false
     }
